re-factored all the code to just have puts on the main.rb and logic in different files
