+++ lib/player.rb
@@ -1,0 +1,7 @@
+class Player
+  attr_reader :name
+
+  def initialize(name)
+    @name = name
+  end
+end
